* dev: (157 commits)
  storage: Change locking in the offset_range_size
  io: add page write accessor
  io: add random sleep helper for tests
  io: add make_page test helper
  io: support custom seed in random data generator
  io: add logger
  rptest: Fix s3.copy_object when running on GCP
  tests: fix lint issue
  slash-cmd: fix /backport command
  c/distributed_kv_stm: prevent invalidated iterators during list
  transform/logging: Tests for log manager
  mock_property: Add call operators
  transform/logging: Introduce transform::logging::manager
  utils: Add sstring comparator to absl_sstring_hash
  transform/logging: Add ss::logger for transform log subsystem
  transform/logging: Introduce transform::logging::client
  config: Transform logging configs
  rpk registry schema get: add flag to print schemas
  tests: restart producer swarm if needed
  k/topics: do not use a vector to wait for leaders
  ...

+++ src/v/model/record.h
@@ -82,7 +82,7 @@ public:
       , _val_size(v_len)
       , _value(std::move(v)) {}
 
-    int32_t memory_usage() const {
+    size_t memory_usage() const {
         return sizeof(*this) + _key.size_bytes() + _value.size_bytes();
     }
     record_header share() {
@@ -167,13 +167,13 @@ public:
 
     // Used for acquiring units from semaphores limiting
     // memory resources.
-    int32_t memory_usage() const {
+    size_t memory_usage() const {
         return sizeof(*this) + _key.size_bytes() + _value.size_bytes()
                + std::accumulate(
                  _headers.begin(),
                  _headers.end(),
-                 int32_t(0),
-                 [](int32_t acc, const record_header& h) {
+                 size_t(0),
+                 [](size_t acc, const record_header& h) {
                      return acc + h.memory_usage();
                  });
     }
@@ -718,7 +718,7 @@ public:
                || (_header.attrs.is_control() && _header.producer_id >= 0);
     }
 
-    int32_t memory_usage() const {
+    size_t memory_usage() const {
         return sizeof(*this) + _records.size_bytes();
     }
